Clear your tidygeoboundaries cache dir

Use this function with caution. This function would clear your cached data and configuration, specifically:

Deletes the tidygeoboundaries config directory (rappdirs::user_config_dir("tidygeobn", "R")).

Deletes the cache_dir directory.

Deletes the values on stored on Sys.getenv("GEOBN_CACHE_DIR") and options(mapSpain_cache_dir).

geobn_clear_cache(config = FALSE, cached_data = TRUE, verbose = FALSE)

Arguments

config

Logical. If TRUE, will delete the configuration folder of tidygeoboundaries.

cached_data

Logical. If TRUE, it will delete your cache_dir and all its content.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

Value

Invisible. This function is called for its side effects.

Details

This is an overkill function that is intended to reset your status as it you would never have installed and/or used tidygeoboundaries.

See also

Examples

# Don't run this! It would modify your current state
if (FALSE) { # \dontrun{
geobn_clear_cache(verbose = TRUE)
} # }

Sys.getenv("GEOBN_CACHE_DIR")
#> [1] ""