Get geoBoundaries data

Return an spatial object using sf.

get_geobn(
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  country = "ALL",
  boundary_type = c("ADM0", "ADM1", "ADM2", "ADM3", "ADM4", "ALL"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL
)

Arguments

release_type

One of "gbOpen", "gbHumanitarian", "gbAuthoritative"`. Source of the spatial data. See Details.

country

A character vector of country codes. It could be either "ALL" (that would return the data for all countries), a vector of country names or a vector of ISO3 country codes. Mixed types (as c("Italy","ES","FRA")) would not work. See also countrycode::countrycode().

boundary_type

Type of boundary Accepted values are "ALL" (all available boundaries) or the ADM level ("ADM0" is the country boundary, "ADM1" is the first level of sub national boundaries, "ADM2" is the second level and so on.

simplified

Logical. Return the simplified boundary or not.

metadata

Should the result be the metadata of the boundary? See Details.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

cache

A logical whether to do caching. Default is TRUE. See About caching.

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source .geojson file.

cache_dir

A path to a cache directory. See geobn_set_cache_dir().

Value

When metadata = FALSE: A sf object, otherwise a tibble.

Details

TODO....

References

geoboundaries API Service https://www.geoboundaries.org/api.html.

Examples

# Metadata
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
get_geobn(country = "Spain", boundary_type = "ADM0", metadata = TRUE) %>%
  glimpse()
#> Rows: 1
#> Columns: 32
#> $ boundaryID                <chr> "ESP-ADM0-43220588"
#> $ boundaryName              <chr> "Spain"
#> $ boundaryISO               <chr> "ESP"
#> $ boundaryYearRepresented   <chr> "2017"
#> $ boundaryType              <chr> "ADM0"
#> $ boundaryCanonical         <chr> "Unknown"
#> $ boundarySource            <chr> "Instituto de Geografico Nacional"
#> $ boundaryLicense           <chr> "Creative Commons Attribution 4.0 Internatio…
#> $ licenseDetail             <chr> "CC-BY 4.0"
#> $ licenseSource             <chr> "centrodedescargas.cnig.es/CentroDescargas/i…
#> $ boundarySourceURL         <chr> "centrodedescargas.cnig.es/CentroDescargas/i…
#> $ sourceDataUpdateDate      <chr> "Thu Jan 19 07:31:04 2023"
#> $ buildDate                 <chr> "Dec 12, 2023"
#> $ Continent                 <chr> "Europe"
#> $ `UNSDG-region`            <chr> "Europe and Northern America"
#> $ `UNSDG-subregion`         <chr> "Southern Europe"
#> $ worldBankIncomeGroup      <chr> "High-income Countries"
#> $ admUnitCount              <chr> "1"
#> $ meanVertices              <chr> "667002.0"
#> $ minVertices               <chr> "667002"
#> $ maxVertices               <chr> "667002"
#> $ meanPerimeterLengthKM     <chr> "13154.542240986726"
#> $ minPerimeterLengthKM      <chr> "13154.542240986726"
#> $ maxPerimeterLengthKM      <chr> "13154.542240986726"
#> $ meanAreaSqKM              <chr> "506009.4752517802"
#> $ minAreaSqKM               <chr> "506009.4752517802"
#> $ maxAreaSqKM               <chr> "506009.4752517802"
#> $ staticDownloadLink        <chr> "https://github.com/wmgeolab/geoBoundaries/r…
#> $ gjDownloadURL             <chr> "https://github.com/wmgeolab/geoBoundaries/r…
#> $ tjDownloadURL             <chr> "https://github.com/wmgeolab/geoBoundaries/r…
#> $ imagePreview              <chr> "https://github.com/wmgeolab/geoBoundaries/r…
#> $ simplifiedGeometryGeoJSON <chr> "https://github.com/wmgeolab/geoBoundaries/r…